Hill cipher in python
Introduction
the Hill cipher is a polygraphic substitution cipher based on linear algebra. In-
vented by Lester S. Hill in 1929, it was the first polygraphic cipher in which it
was practical (though barely) to operate on more than three symbols at once.
The following discussion assumes an elementary knowledge of matrices.

we use python as a programming language and we create 3
matrix because the Hill cipher depends on the matrix and how to deal with
them keyMatrix, MessageVector, cipherMatrix, decipherMatrix. we use Key
matrix to but the key that we generate and use to cipher the plain text and
it 3*3 matrix, MessageVector is the matrix that we represent the plain text
on 3 elements in one matrix to make the multiplication with the key and the
cipherMatrix was the equivalent of this operation, we ll the matrix using a loop
and represent each letter to number with this operation char[i]-65 show
the gure 1 and gure 2 after that we multiple the matrix with each other and
then convert the multiple to text using this equation char[i]+65 show gure 2
show the result in gure 3, if we want to decipher the plain text we only need
to find the inverse matrix and multiple with the cipher matrix